main_app/requirements.txt
fastapi
uvicorn
prometheus-fastapi-instrumentator
requests
kafka-python-ng